+++ src/actions/search/setIsSearchFieldActive.js
@@ -1,0 +1,8 @@
+export default (isActive) => (dispatch) => new Promise((resolve) => {
+    dispatch({
+        type: 'SET_SEARCH_FIELD_ACTIVE',
+        isActive
+    });
+
+    resolve();
+});