+++ src/components/stages/header/WishlistIcon.jsx
@@ -2,10 +2,14 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { MODULE_ID_WISHLIST } from '../../../constants';
 
-const WishlistIcon = ({setActiveModule, translations, wishlistHasItems}) => {
+const WishlistIcon = ({ renderIcon, setActiveModule, translations, wishlistHasItems }) => {
+    if (!renderIcon) {
+        return null;
+    }
+
     return (
-        <div className={'wishlist'} onClick={() => { setActiveModule(MODULE_ID_WISHLIST); }}>
-            <img className={'wishlist__icon'}
+        <div className={'prometheus-header__wishlist'} onClick={() => { setActiveModule(MODULE_ID_WISHLIST); }}>
+            <img className={'prometheus-header__wishlistIcon'}
                  alt={translations.wishlistIconLabel}
                  src={`/images/wishlist-${wishlistHasItems ? 'selected' : 'neutral'}.svg`}
             />
@@ -18,6 +22,7 @@ WishlistIcon.defaultProps = {
 };
 
 WishlistIcon.propTypes = {
+    renderIcon: PropTypes.bool.isRequired,
     setActiveModule: PropTypes.func.isRequired,
     translations: PropTypes.shape({
         wishlistIconLabel: PropTypes.string.isRequired