+++ src/reducers/reducers.js
@@ -2,6 +2,7 @@ import { combineReducers } from 'redux';
 import activeModule from './activeModule';
 import informationItems from './informationItems';
 import navigation from './navigation';
+import search from './search';
 import selectedItemId from './selectedItemId';
 import translations from './translations';
 import user from './user';
@@ -11,6 +12,7 @@ export default combineReducers({
     activeModule,
     informationItems,
     navigation,
+    search,
     selectedItemId,
     translations,
     user,